Stop pytest-asyncio from closing the loop
