--- dask_ctl/tests/test_snippet.py
@@ -1,16 +1,0 @@
-import ast
-
-from dask.distributed import LocalCluster
-
-from dask_ctl.lifecycle import get_snippet
-
-
-def test_snippet():
-    with LocalCluster(scheduler_port=8786) as _:
-        snippet = get_snippet("proxycluster-8786")
-
-        # Check is valid Python
-        ast.parse(snippet)
-
-        assert "ProxyCluster" in snippet
-        assert "proxycluster-8786" in snippet
